+++ src/components/TableHeader/index.tsx
@@ -1,0 +1,19 @@
+import { Table } from "@mantine/core";
+import { TableProps } from "../../type/type";
+import { getHeader } from "../../util/header.util";
+
+import TableHeaderRow from "./TableHeaderRow";
+
+const TableHeader = <T,>({ table, headerOptionType }: TableProps<T>) => {
+  const headerGroups = getHeader({ table, headerOptionType });
+
+  return (
+    <Table.Thead>
+      {headerGroups.map((headerGroup) => (
+        <TableHeaderRow key={headerGroup.depth} headerGroup={headerGroup} />
+      ))}
+    </Table.Thead>
+  );
+};
+
+export default TableHeader;
