+++ src/util/header.util.ts
@@ -1,12 +1,14 @@
 import { Header } from "@tanstack/react-table";
 import { TableProps } from "../type/type";
 
-interface HeaderDataType<T> {
+export interface HeaderDataType<T> {
   depth: number;
   headers: Header<T, unknown>[];
 }
 
-const groupByDepthAndSort = <T>(array: Header<T, unknown>[]): HeaderDataType<T>[] => {
+const groupByDepthAndSort = <T>(
+  array: Header<T, unknown>[]
+): HeaderDataType<T>[] => {
   // Step 1: depth에 따라 그룹화
   const grouped = array.reduce<HeaderDataType<T>[]>((acc, obj) => {
     const existingGroup = acc.find((group) => group.depth === obj.depth);
